src/requirements.txt
blinker==1.7.0
Flask==3.0.2
itsdangerous==2.1.2
psycopg2-binary==2.9.9
Werkzeug==3.0.1
